Fix Status fields returned in certain response commands.
Note that Lock, Unlock, Toggle, and UnlockWithTimeout define an enum8 status field containing ZCL status values SUCCESS and FAILURE (0 and 1). These are annotated to refer to the zclStatus type.

Many of the rest of the status fields in response commands are specified to be of type uint8 with values Pass and Fail (0 and 1). Status type DrlkPassFailStatus is defined for these.

Another status type DrlkSetCodeStatus is defined for SetPIN and SetRFID response commands. This status reports Success, GeneralFailure, MemoryFull, and DuplicateCode (the latter three considered errors).