Add implementation dependencies for optional attributes.
TempThreshold and its corresponding TempDwellTripPoint are mutually dependent.
DeviceTempAlarmMask must be implemented if either Low or High TempThreshold is implemented. It is used to enable High and Low alarms, its default is zero (both disabled), and the entire point of this attribute set is to support alarming, so it makes no sense to implement any of the other attributes but not this one.

Note, the implementation dependencies for these optional attributes are implied by text that describes behavior or interactions among the attributes; but those dependencies are not stated explicitly in the specification text.