Add placeholder entries in CHANGELOG for all ZCL files expected to be updated for ZCL8
Added headings (alphabetical by XML filename) for all of the clusters having a ZCL8 Trello card. I believe this will reduce the number of merge conflicts as this file is updated.

Marked a few of these entries as NEW if the corresponding XML file does not already exist (in develop branch).